Created seperate order card component in client side
